# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Prisma Embedded Codes — Authentication & Authorization System API
  description: Security-hardened REST API for user registration, authentication, MFA (TOTP/Email), token rotation, and RBAC authorization.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
    description: Local Development Server
paths:
...
  /users/me:
    get:
      summary: Retrieve current profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  email:
                    type: string
                  fullName:
                    type: string
                  role:
                    type: string
  /users/admin/users:
    get:
      summary: Retrieve user list (Admin only)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
...